test: overhaul test infrastructure, fill coverage gaps, fix event contract
Split agent/tests/test_unit.py (1097 lines, 48 tests) into 9 focused modules
(88 tests total, +40 new). Parametrize duplicated test patterns. Add shared
fixtures to conftest.py.

New coverage:
- test_notifications.py: notification loading, parsing, deletion, batching,
  process_batch, _is_new_json filter (previously zero unit tests)
- test_contract.py: verify Python TypedDicts match TypeScript VestaEvent union,
  JSON serialization roundtrip, EventBus persistence for all event types
- types.test.ts: compile-time contract test for all 13 event types
- CLI unit tests: format_size, normalize_url, version_less_than

Fixes:
- Add missing subagent_start/subagent_stop to TypeScript VestaEvent type
- Add missing subagent field to tool_start/tool_end in TypeScript types

CI:
- Wire tasks skill CLI tests into pr-checks (self-contained, no creds needed)
- Run all test modules instead of just test_unit.py
