+++ .github/workflows/pr-checks.yml
@@ -126,4 +126,8 @@ jobs:
 
       - name: Run tests
         working-directory: agent
-        run: uv run pytest tests/test_unit.py
+        run: uv run pytest tests/ --ignore=tests/test_e2e.py -v
+
+      - name: Run skill CLI tests
+        working-directory: agent
+        run: uv run pytest skills/tasks/cli/tests/ -v
