+++ agent/tests/test_dreamer.py
@@ -1,0 +1,81 @@
+"""Tests for nightly dreamer/memory scheduling."""
+
+import asyncio
+import datetime as dt
+from unittest.mock import patch
+
+import pytest
+import vesta.models as vm
+from test_processor import _run_processor_test
+
+
+@pytest.mark.anyio
+async def test_queues_prompt_and_archives(tmp_path):
+    from vesta.core.loops import process_nightly_memory
+
+    state = vm.State()
+    state.last_dreamer_run = None
+    queue: asyncio.Queue = asyncio.Queue()
+
+    dreamer_hour = 4
+    config = vm.VestaConfig(root=tmp_path, nightly_memory_hour=dreamer_hour)
+    fake_now = dt.datetime(2025, 6, 15, dreamer_hour, 0, 0)
+
+    with (
+        patch("vesta.core.loops._now", return_value=fake_now),
+        patch("vesta.core.loops.load_prompt", return_value="dreamer prompt"),
+    ):
+        await process_nightly_memory(queue, state=state, config=config)
+
+    assert not queue.empty()
+    msg, is_user = await queue.get()
+    assert msg == "dreamer prompt"
+    assert is_user is False
+    assert state.last_dreamer_run == fake_now
+    assert state.dreamer_active is True
+
+
+@pytest.mark.anyio
+async def test_skips_when_already_run_today(tmp_path):
+    from vesta.core.loops import process_nightly_memory
+
+    dreamer_hour = 4
+    config = vm.VestaConfig(root=tmp_path, nightly_memory_hour=dreamer_hour)
+    fake_now = dt.datetime(2025, 6, 15, dreamer_hour, 0, 0)
+
+    state = vm.State()
+    state.last_dreamer_run = fake_now
+    queue: asyncio.Queue = asyncio.Queue()
+
+    with (
+        patch("vesta.core.loops._now", return_value=fake_now),
+        patch("vesta.core.loops.load_prompt", return_value="dreamer prompt"),
+    ):
+        await process_nightly_memory(queue, state=state, config=config)
+
+    assert queue.empty()
+    assert state.dreamer_active is False
+
+
+@pytest.mark.anyio
+async def test_compacts_and_restarts(tmp_path):
+    async def side_effect(msg, *, state, config, is_user):
+        return (["OK"], state)
+
+    pre_state = vm.State()
+    pre_state.session_id = "pre-dreamer-session"
+    pre_state.dreamer_active = True
+
+    fake_now = dt.datetime(2025, 6, 15, 4, 5, 0)
+    state, session_count, messages = await _run_processor_test(
+        tmp_path,
+        message_side_effect=side_effect,
+        pre_state=pre_state,
+        initial_queue=[("dreamer prompt content", False)],
+        extra_patches={"vesta.core.loops._now": lambda: fake_now},
+    )
+    assert state.session_id == "pre-dreamer-session"
+    assert state.dreamer_active is False
+    assert state.graceful_shutdown.is_set()
+    assert state.restart_reason == "nightly — dreamer ran, context compacted"
+    assert messages == ["dreamer prompt content", "/compact"]
